+++ GameServer/AscensionServer/Model/GongFa.cs
@@ -15,7 +15,7 @@ namespace AscensionServer.Model
     public class GongFa:Model
     {
         public virtual int ID { get; set; }
-        public virtual byte GongFaID { get; set; }
+        public virtual int GongFaID { get; set; }
         public virtual int GongFaExp { get; set; }
         public virtual short GongFaLevel { get; set; }
         public virtual string GongFaSkillArray { get; set; }
