+++ GameServer/AscensionServer/Model/MiShu.cs
@@ -15,7 +15,7 @@ namespace AscensionServer.Model
     public class MiShu:Model
     {
         public virtual int ID { get; set; }
-        public virtual byte MiShuID { get; set; }
+        public virtual int MiShuID { get; set; }
         public virtual int MiShuExp { get; set; }
         public virtual short MiShuLevel { get; set; }
         public virtual string MiShuSkillArry { get; set; }
